+++ catclass/__init__.py
@@ -1,1 +1,2 @@
 from .catclass import CatClassifier
+from .cli import CLI